requirements.txt
matplotlib==2.2.3
torchvision==0.4.2
opencv_python==4.1.2.30
requests==2.24.0
numpy==1.19.1
torch==1.3.1
web.py==0.61
Pillow==7.2.0
web==0.6.0
